Move Toucan cleanup to standardization
